+++ src/components/CommentForm/index.js
@@ -1,0 +1,3 @@
+import component from './component'
+import connector from './connector'
+export default connector(component)
